Reference commit: 99.82% accuracy on test set.
Added random crops to the train-time augmentation, changed
the crop size slightly, and at test time perform a single
crop of the same size, centered.  In all cases resized back
to 28x28.

The single one-cycle with ten epochs total results in classifiers
that typically have accuracies in the mid 99.70s.  Then committee
softmax voting pushes it to the high 99.70s or low 99.80's, depending
on the committee size.

Also experimented with test-time augumentation, but it is not currently
used.

Code is slightly rough at the moment.

$ python mnist.py
1: Softmax committee of 15 accuracy: 99.82%
2: Softmax committee of 15 accuracy: 99.81%
3: Softmax committee of 15 accuracy: 99.80%
4: Softmax committee of 15 accuracy: 99.79%
5: Softmax committee of 15 accuracy: 99.79%
6: Softmax committee of 15 accuracy: 99.79%
7: Softmax committee of 15 accuracy: 99.80%
8: Softmax committee of 15 accuracy: 99.79%
9: Softmax committee of 15 accuracy: 99.82%
10: Softmax committee of 15 accuracy: 99.78%
11: Softmax committee of 15 accuracy: 99.81%
12: Softmax committee of 15 accuracy: 99.79%
13: Softmax committee of 15 accuracy: 99.81%
14: Softmax committee of 15 accuracy: 99.83%
15: Softmax committee of 15 accuracy: 99.78%
mean: 99.80% 0.9980066666666667

$ python mnist.py
1: Softmax committee of 35 accuracy: 99.82%