Fixed subtle bug.
A recent change to Classifier apparently had subtly impaired the training performance,
and I believe the reason is that a required .detach() was removed.  This probably
caused the status reports during training to slightly impair subsequent gradient
calculations.  I thought that doing the operation in the context of no_grad()
was sufficient, but apparently this is not the case.

I also separated out some code for running a particular experiment from
the main code.